+++ Decsys/ClientApp/src/app/screens/survey/SurveyScreen.js
@@ -3,7 +3,7 @@ import { useNavigation } from "react-navi";
 import * as api from "api";
 import SurveyPage from "components/SurveyPage";
 import AppBar from "components/AppBar";
-import { SURVEY_COMPLETE } from "utils/event-types";
+import { SURVEY_COMPLETE } from "constants/event-types";
 import AboutLink from "components/AboutLink";
 
 // TODO: PropTypes
