+++ Decsys/ClientApp/src/components/SurveyPage/SurveyPage.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
 import { Grid, Cell } from "styled-css-grid";
-import { FlexBox, Container } from "../../components/ui";
+import { FlexBox, Container } from "components/ui";
 import { Button } from "@smooth-ui/core-sc";
 import { ChevronRight } from "styled-icons/fa-solid";
-import { PAGE_LOAD } from "../../utils/event-types";
+import { PAGE_LOAD } from "constants/event-types";
 import SurveyPageBody from "./Body";
-import { getResponseComponent } from "../../utils/component-utils";
+import { getResponseComponent } from "utils/component-utils";
 
 // TODO: Prop Types!
