+++ Decsys/ClientApp/src/constants/event-types.js
